Move Span mark to bar module
