+++ seaborn/_marks/bar.py
@@ -248,3 +248,70 @@ class Bars(BarBase):
             linewidth = min(.1 * min_width, mpl.rcParams["patch.linewidth"])
             for _, col in collections.items():
                 col.set_linewidth(linewidth)
+
+
+@document_properties
+@dataclass
+class Span(BarBase):
+    """
+    TODO
+    """
+    color: MappableColor = Mappable("C0")
+    alpha: MappableFloat = Mappable(.7)
+    fill: MappableBool = Mappable(True)
+    edgecolor: MappableColor = Mappable(depend="color")
+    edgealpha: MappableFloat = Mappable(1)
+    edgewidth: MappableFloat = Mappable(rc="patch.linewidth")
+    edgestyle: MappableFloat = Mappable("-")
+    width: MappableFloat = Mappable(0.8, grouping=False)
+
+    def _plot(self, split_gen, scales, orient):
+
+        patches = defaultdict(list)
+
+        for keys, data, ax in split_gen():
+
+            kws = {}
+
+            resolved = resolve_properties(self, keys, scales)
+            fc = resolve_color(self, keys, "", scales)
+            if not resolved["fill"]:
+                fc = mpl.colors.to_rgba(fc, 0)
+
+            kws["facecolor"] = fc
+            kws["edgecolor"] = resolve_color(self, keys, "edge", scales)
+            kws["linewidth"] = resolved["edgewidth"]
+            kws["linestyle"] = resolved["edgestyle"]
+
+            other = {"x": "y", "y": "x"}[orient]
+
+            if not set(data.columns) & {f"{other}min", f"{other}max"}:
+                agg = {f"{other}min": (other, "min"), f"{other}max": (other, "max")}
+                data = data.groupby([orient, "width"]).agg(**agg).reset_index()
+
+            for row in data.itertuples():
+
+                if orient == "x":
+
+                    verts = np.array([
+                        row.x - row.width / 2, row.ymin,
+                        row.x + row.width / 2, row.ymin,
+                        row.x + row.width / 2, row.ymax,
+                        row.x - row.width / 2, row.ymax,
+                    ]).reshape((4, 2))
+                else:
+                    verts = np.array([
+                        row.xmin, row.y - row.width / 2,
+                        row.xmax, row.y - row.width / 2,
+                        row.xmax, row.y + row.width / 2,
+                        row.xmin, row.y + row.width / 2,
+                    ]).reshape((4, 2))
+
+                patches[ax].append(mpl.patches.Polygon(verts, **kws))
+                ax.update_datalim(verts)
+
+        for ax, ax_patches in patches.items():
+
+            for patch in ax_patches:
+                self._postprocess_artist(patch, ax, orient)
+                ax.add_patch(patch)
