+++ seaborn/objects.py
@@ -29,8 +29,8 @@ ways that a plot can be enhanced and customized.
 from seaborn._core.plot import Plot  # noqa: F401
 
 from seaborn._marks.base import Mark  # noqa: F401
-from seaborn._marks.area import Area, Band, Span  # noqa: F401
-from seaborn._marks.bar import Bar, Bars  # noqa: F401
+from seaborn._marks.area import Area, Band  # noqa: F401
+from seaborn._marks.bar import Bar, Bars, Span  # noqa: F401
 from seaborn._marks.dot import Dot, Dots  # noqa: F401
 from seaborn._marks.line import Dash, Line, Lines, Path, Paths, Range  # noqa: F401
 from seaborn._marks.text import Text  # noqa: F401
